src/rfx/requirements.txt
filelock
urllib3
ujson
pyparsing
awscli
python-magic
pynacl
requests
pylint
pymongo
cherrypy
setproctitle
slackclient
dictlib
onetimejwt
timeinterval
